user registraiton with verification
